requirements.txt
tf_keras
requests
Flask
Werkzeug~=3.0.4
Authlib~=1.3.2
Flask==2.3.2
Werkzeug
skitlearn
tensorflow
pandas
python-dotenv
Flask-cors
scikit-learn
numpy
Pillow
gunicorn  # Optional for production
pdfkit
scipy

